requirements.txt
# RAID Manager Dependencies

# Core
pyyaml>=6.0

# Desktop UI (install when building GUI)
# customtkinter>=5.2

# CLI enhancements (optional)
# click>=8.0

# For building standalone executables
# pyinstaller>=6.0
